Module aws_sdk_iot1clickprojects::operation

Structs

AssociateDeviceWithPlacement

Associates a physical device with a placement.

CreatePlacement

Creates an empty placement.

CreateProject

Creates an empty project with a placement template. A project contains zero or more

DeletePlacement

Deletes a placement. To delete a placement, it must not have any devices associated with

DeleteProject

Deletes a project. To delete a project, it must not have any placements associated with

DescribePlacement

Describes a placement in a project.

DescribeProject

Returns an object describing a project.

DisassociateDeviceFromPlacement

Removes a physical device from a placement.

GetDevicesInPlacement

Returns an object enumerating the devices in a placement.

ListPlacements

Lists the placement(s) of a project.

ListProjects

Lists the AWS IoT 1-Click project(s) associated with your AWS account and region.

ListTagsForResource

Lists the tags (metadata key/value pairs) which you have assigned to the resource.

TagResource

Creates or modifies tags for a resource. Tags are key/value pairs (metadata) that can be

UntagResource

Removes one or more tags (metadata key/value pairs) from a resource.

UpdatePlacement

Updates a placement with the given attributes. To clear an attribute, pass an empty value

UpdateProject

Updates a project associated with your AWS account and region. With the exception of